revision ID 타입 변경
이력 ID 고갈 문제가 있을 수 있음
Integer -> Long

+++ src/main/java/com/megazone/bomdi/cms/repository/ContractRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.megazone.bomdi.cms.domain.Contract;
 import org.springframework.data.repository.history.RevisionRepository;
 
-public interface ContractRepository extends JpaRepository<Contract, Long>, RevisionRepository<Contract, Long, Integer> {
+public interface ContractRepository extends JpaRepository<Contract, Long>, RevisionRepository<Contract, Long, Long> {
 }
